BirthdayCalendar renaming class
